software: Split file into modules
